% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSplice-designMatrix.R
\docType{data}
\name{designMatrix}
\alias{designMatrix}
\title{An example of a design matrix helpful to demonstrate
the use of the NBSplice R package.}
\format{A data.frame object}
\source{
see \code{\link{IsoDataSet-class}}
}
\description{
A data.frame representing the design matrix related to the isoCounts dataset.
}
\details{
\describe{
Each row of this data.frame specifies one experiment sample.
\item{samples}{Column with samples names}.
\item{condition}{Column indicating the condition of each sample}.
}
}
\seealso{
Other IsoDataSet: \code{\link{IsoDataSet-class}},
  \code{\link{IsoDataSet}}, \code{\link{NBTest}},
  \code{\link{buildData}}, \code{\link{buildLowExpIdx}},
  \code{\link{geneIso}}, \code{\link{initialize}},
  \code{\link{isoCounts}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and 
Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
