% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSplice-geneIso.R
\docType{data}
\name{geneIso}
\alias{geneIso}
\title{An example of isoform-gene relationship specification helpful to demonstrate
the use of the NBSplice R package.}
\format{A data.frame object}
\source{
see \code{\link{IsoDataSet-class}}
}
\description{
A dataset containing isoform-gene relationships.
}
\details{
\describe{
Each row of this data.frame specifies one specific and unique isoform-gene
relationship.
\item{gene_id}{Column with the gene names}.
\item{isoform_id}{Column with the isoform names}.
}
}
\seealso{
Other IsoDataSet: \code{\link{IsoDataSet-class}},
  \code{\link{IsoDataSet}}, \code{\link{NBTest}},
  \code{\link{buildData}}, \code{\link{buildLowExpIdx}},
  \code{\link{designMatrix}}, \code{\link{initialize}},
  \code{\link{isoCounts}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and 
Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
