% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSplice-isoCounts.R
\docType{data}
\name{isoCountsData}
\alias{isoCountsData}
\title{An example of an isoform counts matrix helpful to demonstrate the use of
the NBSplice R package.}
\format{A data.frame object}
\source{
see \code{\link{IsoDataSet-class}}
}
\description{
A non-real dataset containing isoform expression counts of a synthetic 
dataset where several differential splicing changes were simulated and
controlled.
}
\details{
\describe{
Isoforms are represented in rows and samples in columns. Samples 'C1R1', 
'C1R2', 'C1R3' and 'C1R4' are from Normal condition whereas the other samples
are from Tumor condition.
}
}
\seealso{
Other IsoDataSet: \code{\link{IsoDataSet-class}},
    \code{\link{IsoDataSet}}, \code{\link{NBTest}},
    \code{\link{buildData}}, \code{\link{buildLowExpIdx}},
    \code{\link{designMatrix}}, \code{\link{geneIso}},
    \code{\link{initialize}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and 
Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
