% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet-buildLowExpIdx.R
\docType{methods}
\name{buildLowExpIdx}
\alias{buildLowExpIdx}
\alias{buildLowExpIdx-methods}
\alias{buildLowExpIdx}
\alias{buildLowExpIdx,IsoDataSet-method}
\title{IsoDataSet low expressed isoforms detection.}
\usage{
buildLowExpIdx(object, colName = "condition", ratioThres = 0.01,
  countThres = 1, BPPARAM = bpparam())

\S4method{buildLowExpIdx}{IsoDataSet}(object, colName = "condition",
  ratioThres = 0.01, countThres = 1, BPPARAM = bpparam())
}
\arguments{
\item{object}{IsoDataSet class object.}

\item{colName}{Character indicating the name of the column in the design 
matrix to be considered for differential expression analysis.}

\item{ratioThres}{Numeric indicating the minimum isoform's relative expression
value admitted. If one isoform had expression lower than this threshold in at
least one sample, thus it will be ignored for further analysis.}

\item{countThres}{Numeric indicating the isoform's expression threshold. If 
one isoform showed a mean expression value lower than this threshold in at
least one experimental condition, thus it will be ignored for further
analysis.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
IsoDataSet object.
}
\description{
\code{buildLowExpIdx} identifies low expressed isoforms and 
stores their indexes to facilitate their identification. Low expression is 
defined  combining absolute and relative expression thresholds.
}
\note{
see full example in \code{\link{IsoDataSet-class}}
}
\examples{
## Data loading
data(myIsoDataSet, package="NBSplice")

## Identification of low expressed isoforms
myIsoDataSet<-buildLowExpIdx(myIsoDataSet)
}
\seealso{
\code{\link{IsoDataSet}}

Other IsoDataSet: \code{\link{IsoDataSet-class}},
  \code{\link{IsoDataSet}}, \code{\link{NBTest}},
  \code{\link{buildData}}, \code{\link{designMatrix}},
  \code{\link{geneIso}}, \code{\link{initialize}},
  \code{\link{isoCounts}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
