% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet.R
\docType{class}
\name{IsoDataSet-class}
\alias{IsoDataSet-class}
\title{IsoDataSet S4 class implementation in R}
\description{
This S4 class represents a data set containing isoforms expresion in R.
}
\section{Slots}{

\describe{
\item{\code{counts}}{Matrix containing expression values at the isoform level}

\item{\code{geneCounts}}{Matrix with expression values at the gene level}

\item{\code{colData}}{Data.frame with experiment information. Its rows should 
be the columns of the counts data.frame}

\item{\code{isoGeneRel}}{Data.frame specifying the isoform-gene relationship}

\item{\code{design}}{Formula to be used in the GLM fit and tests}

\item{\code{lowExpIndex}}{Numeric indicating the positions of low expressed 
isoforms}
}}

\section{Features}{

\enumerate{
    \item Discover differential modifications in the splicing patterns. 
    \item Detect and quantify isoform relative expression changes.
    \item Combine the results of both gene and isoform levels analysis.
}
}

\section{Functions}{

IsoDataSet S4 class includes the following functions:
\describe{
\item{initialize}{Constructor of IsoDataSet objects.} 
\item{isoCounts}{Get the counts slot.}
\item{geneCounts}{Get the geneCounts slot.}
\item{isoGeneRel}{Get the isoGeneRel slot.}
\item{colData}{Get the colData slot.}
\item{designFormula}{Get the design slot.}
\item{lowExpIndex}{Get the lowExpIdx slot.}
\item{setDesign}{Set the design slot.}
\item{buildLowExpIdx}{Build the index to identify low expressed isoforms.}
\item{NBTest}{Perform differential expression analysis.}
}
}

\examples{
## Data loading
data(isoCounts, package="NBSplice")
data(designMatrix, package="NBSplice")
data(geneIso, package="NBSplice")

## Arguments definition
colName<-"condition"

## Constructor calling
myIsoDataSet<-IsoDataSet(isoCounts, designMatrix, colName, geneIso)

## Identificating Low expressed Isoforms
myIsoDataSet<-buildLowExpIdx(myIsoDataSet)

##Arguments definition
colName<-"condition"
test<-"F"

## Differential splicing test
myDSResults<-NBTest(myIsoDataSet, colName, test)
}
\seealso{
Other IsoDataSet: \code{\link{IsoDataSet}},
    \code{\link{NBTest}}, \code{\link{buildData}},
    \code{\link{buildLowExpIdx}}, \code{\link{designMatrix}},
    \code{\link{geneIso}}, \code{\link{initialize}},
    \code{\link{isoCountsData}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A.
Fernandez \email{efernandez@bdmg.com.ar}
}
