% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet-core.R
\docType{methods}
\name{buildData}
\alias{buildData}
\alias{buildData-methods}
\alias{fitModel}
\alias{fitModel-methods}
\title{GLM fitting and hypothesis testing}
\usage{
buildData(isoCounts, geneCounts, geneIso, gene, designMatrix, colName)

fitModel(myData, gene, formula, colName, test = c("F", "Chisq"), contrast)
}
\arguments{
\item{isoCounts}{Matrix having the expression counts at the isoform level. 
Isoforms must be in rows and samples in columns. Rownames and colnames must 
be defined with isoform and samples names, respectively.}

\item{geneCounts}{Matrix having the expression counts at the gene level. 
Genes must be in rows and samples in columns. Rownames and colnames must 
be defined with gene and sample names, respectively.}

\item{geneIso}{Data.frame containing the relationship between isoforms and 
genes. It must contain two columns, named as 'gene_id' and 'isoform_id'.
Its isoforms should be the same specified in the isoCounts matrix.}

\item{gene}{Character indicating the name of the gene to be analyzed}

\item{designMatrix}{Data.frame specifying metadata related to the
experiment. Its rows must be the samples and experimental factors should be
arranged on its columns.}

\item{colName}{Character indicating the name of the column in the design 
matrix to be considered for mean expression calculations per experimental
condition and differential expression test.}

\item{myData}{Data.frame containing the expression matrix at isoform 
levels, it means, isoform in rows and samples in columns. It is obtained 
using the buildData NBSplice method.}

\item{formula}{Object with the formula of the GLM.}

\item{test}{Character indicating the name of the distribution to assume for
linear hypothesis statistic. Could be "F" or "chisq".}

\item{contrast}{Character vector with the names of the two levels of the
experimental factor to be contrasted.}
}
\value{
A data.frame ready to use by the model fitting function

A data.frame summarizing gene results.
}
\description{
\code{buildData} is an R function to build the data
structure useful to fit a negative binomial model at the gene level.
\code{fitModel} is an R function to fit a negative binomial generalized
linear model in order to evaluate changes in the isoform ratio. Each isoform
is considered as a factor. The model incorporates one factor related to one
experimental condition with two levels.
}
\note{
see full example in \code{\link{IsoDataSet-class}}
}
\examples{
## Data loading
data(myIsoDataSet, package="NBSplice")

## Identificating Low expressed Isoforms
myIsoDataSet<-buildLowExpIdx(myIsoDataSet)

##Arguments definition
colName<-"condition"
test<-"F"

## Differential splicing test
myDSResults<-NBTest(myIsoDataSet, colName, test)

}
\seealso{
\code{\link{IsoDataSet}}

Other IsoDataSet: \code{\link{IsoDataSet-class}},
    \code{\link{IsoDataSet}}, \code{\link{NBTest}},
    \code{\link{buildLowExpIdx}}, \code{\link{designMatrix}},
    \code{\link{geneIso}}, \code{\link{initialize}},
    \code{\link{isoCountsData}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A.
Fernandez \email{efernandez@bdmg.com.ar}

Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A.
Fernandez \email{efernandez@bdmg.com.ar}
## Data loading
data(myIsoDataSet, package="NBSplice")

## Identificating Low expressed Isoforms
myIsoDataSet<-buildLowExpIdx(myIsoDataSet)

##Arguments definition
colName<-"condition"
test<-"F"

## Differential splicing test
myDSResults<-NBTest(myIsoDataSet, colName, test)
}
