% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet-getters.R
\docType{methods}
\name{IsoDataSet-getters}
\alias{IsoDataSet-getters}
\alias{isoCounts}
\alias{isoCounts-methods}
\alias{isoCounts}
\alias{isoCounts,IsoDataSet-method}
\alias{geneCounts}
\alias{geneCounts-methods}
\alias{geneCounts}
\alias{geneCounts,IsoDataSet-method}
\alias{isoGeneRel}
\alias{isoGeneRel-methods}
\alias{isoGeneRel}
\alias{isoGeneRel,IsoDataSet-method}
\alias{expData}
\alias{expData-methods}
\alias{expData}
\alias{expData,IsoDataSet-method}
\alias{designFormula}
\alias{designFormula-methods}
\alias{designFormula}
\alias{designFormula,IsoDataSet-method}
\alias{lowExpIndex}
\alias{lowExpIndex-methods}
\alias{lowExpIndex}
\alias{lowExpIndex,IsoDataSet-method}
\title{Getters for an IsoDataSet object.}
\usage{
isoCounts(object, CPM = TRUE)

\S4method{isoCounts}{IsoDataSet}(object, CPM = TRUE)

geneCounts(object)

\S4method{geneCounts}{IsoDataSet}(object)

isoGeneRel(object)

\S4method{isoGeneRel}{IsoDataSet}(object)

expData(object)

\S4method{expData}{IsoDataSet}(object)

designFormula(object)

\S4method{designFormula}{IsoDataSet}(object)

lowExpIndex(object)

\S4method{lowExpIndex}{IsoDataSet}(object)
}
\arguments{
\item{object}{IsoDataSet class object.}

\item{CPM}{Logical indicating if expression matrix at the isoform level 
should be in CPM scale.}
}
\value{
According to the call one of the following objects can be returned
\item{matrix}{Isoform expression matrix.}
\item{matrix}{Gene expression matrix.}
\item{data.frame}{Experiment information.}
\item{data.frame}{Isoforms-gene relationships.}
\item{formula}{Formula to be used in the GLM fit.}
\item{numeric}{Index of low expressed isoforms.}
}
\description{
Obtain information about IsoDataSet slots, according to the given
function call.
}
\note{
see full example in \code{\link{IsoDataSet-class}}
}
\examples{
## Data loading
data(myIsoDataSet, package="NBSplice")

## Getting isoform expression counts
myCounts<-isoCounts(myIsoDataSet)
head(myCounts)
## Data loading
data(myIsoDataSet, package="NBSplice")

## Getting gene expression counts
myGeneCounts<-geneCounts(myIsoDataSet)
head(myGeneCounts)
## Data loading
data(myIsoDataSet, package="NBSplice")

## Getting isoform-gene relationships
myIsoGeneRel<-isoGeneRel(myIsoDataSet)
head(myIsoGeneRel)
## Data loading
data(myIsoDataSet, package="NBSplice")

## Getting the design matrix
myDesignMatrix<-expData(myIsoDataSet)
myDesignMatrix
## Data loading
data(myIsoDataSet, package="NBSplice")

## Getting the model formula
myFormula<-designFormula(myIsoDataSet)
myFormula
## Data loading
data(myIsoDataSet, package="NBSplice")

## Getting index of low expressed isoforms
myLowExpIdx<-lowExpIndex(myIsoDataSet)
head(myLowExpIdx)
}
\seealso{
\code{\link{IsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A.
Fernandez \email{efernandez@bdmg.com.ar}
}
