% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet-initialize.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,IsoDataSet-method}
\title{IsoDataSet object constructor.}
\usage{
\S4method{initialize}{IsoDataSet}(.Object, isoCounts, experimentData, colName,
    geneIso, BPPARAM = bpparam())
}
\arguments{
\item{.Object}{IsoDataSet class object.}

\item{isoCounts}{Matrix having the expression counts at the isoform level. 
Isoforms must be in rows and samples in columns. Rownames and colnames must 
be defined with isoform and samples names, respectively.}

\item{experimentData}{Data.frame specifying metadata related to the
experiment. Its rows must be the samples and experimental factors should be
arranged on its columns.}

\item{colName}{Character indicating the name of the column in the design 
matrix to be considered for mean expression calculations per experimental
condition and differential expression test.}

\item{geneIso}{Data.frame containing the relationship between isoforms and 
genes. It must contain two columns, named as 'gene_id' and 'isoform_id'.
Its isoforms should be the same specified in the isoCounts matrix.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
IsoDataSet object.
}
\description{
\code{initialize} creates an IsoDataSet object with the specified isoform
expression counts.
}
\note{
see full example in \code{\link{IsoDataSet-class}}
}
\examples{

## Data loading
data(isoCounts, package="NBSplice")
data(designMatrix, package="NBSplice")
data(geneIso, package="NBSplice")

## Arguments definition
colName<-"condition"

## Constructor calling
myIsoDataSet<-IsoDataSet(isoCounts, designMatrix, colName, geneIso)

}
\seealso{
\code{\link{IsoDataSet}}

Other IsoDataSet: \code{\link{IsoDataSet-class}},
    \code{\link{IsoDataSet}}, \code{\link{NBTest}},
    \code{\link{buildData}}, \code{\link{buildLowExpIdx}},
    \code{\link{designMatrix}}, \code{\link{geneIso}},
    \code{\link{isoCountsData}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
