% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet-NBTest.R
\docType{methods}
\name{NBTest}
\alias{NBTest}
\alias{NBTest-methods}
\alias{NBTest}
\alias{NBTest,IsoDataSet-method}
\title{Differential splicing analysis.}
\usage{
NBTest(object, colName = "condition", test = c("F", "Chisq"),
    contrast = c(levels(expData(object)[, colName])[1:2]),
    BPPARAM = bpparam())

\S4method{NBTest}{IsoDataSet}(object, colName = "condition", test = c("F",
    "Chisq"), contrast = c(levels(expData(object)[, colName])[1:2]),
    BPPARAM = bpparam())
}
\arguments{
\item{object}{IsoDataSet class object.}

\item{colName}{Character indicating the name of the column in the design 
matrix to be considered for mean expression calculations per experimental
condition.}

\item{test}{Character indicating the name of the distribution to assume for
linear hypothesis statistic. Could be "F" or "chisq".}

\item{contrast}{Character vector with the names of the two levels of the
experimental factor to be contrasted.}

\item{BPPARAM}{An optional BiocParallelParam instance defining the parallel
back-end to be used during evaluation.}
}
\value{
A NBSpliceRes object with the obtained results.
}
\description{
\code{NBTest} is the main function of NBSplice. It fits the negative binomial
models for all genes and performs the corresponding hypothesis tests to
evaluate the occurrence of differential splicing.
}
\note{
see full example in \code{\link{IsoDataSet-class}}
}
\examples{
## Data loading
data(myIsoDataSet, package="NBSplice")

## Identificating Low expressed Isoforms
myIsoDataSet<-buildLowExpIdx(myIsoDataSet)

##Arguments definition
colName<-"condition"
test<-"F"

## Differential splicing test
myDSResults<-NBTest(myIsoDataSet, colName, test)

}
\seealso{
\code{\link{IsoDataSet}}

Other IsoDataSet: \code{\link{IsoDataSet-class}},
    \code{\link{IsoDataSet}}, \code{\link{buildData}},
    \code{\link{buildLowExpIdx}}, \code{\link{designMatrix}},
    \code{\link{geneIso}}, \code{\link{initialize}},
    \code{\link{isoCountsData}}, \code{\link{myIsoDataSet}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
