% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet-print.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,IsoDataSet-method}
\title{Print an IsoDataSet object.}
\usage{
\S4method{print}{IsoDataSet}(x, ...)
}
\arguments{
\item{x}{IsoDataSet class object.}

\item{...}{Included for generic print compatibility.}
}
\value{
Console output of the object.
}
\description{
Generic \code{print} method for IsoDataSet class and descendants.
}
\note{
see full example in \code{\link{IsoDataSet-class}}
}
\examples{

## Data loading
data(myIsoDataSet, package="NBSplice")

print(myIsoDataSet)

}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com},
and Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
