% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoDataSet-show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,IsoDataSet-method}
\title{Show method for the IsoDataSet class.}
\usage{
\S4method{show}{IsoDataSet}(object)
}
\arguments{
\item{object}{IsoDataSet class object}
}
\value{
Console output of the object
}
\description{
\code{show} an IsoDataSet object
}
\note{
see full example in \code{\link{IsoDataSet-class}}
}
\examples{

## Data loading
data(myIsoDataSet, package="NBSplice")

show(myIsoDataSet)

}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com}
and Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
