% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSplice-myDSResults.R
\docType{data}
\name{myDSResults}
\alias{myDSResults}
\title{An example of an NBSpliceRes object.}
\format{A NBSpliceRes object}
\source{
see \code{\link{NBSpliceRes-class}}
}
\description{
An example of an NBSpliceRes object helpful to demonstrate the use of
the NBSplice R package.

\describe{
An NBSpliceRes object built obtained using the code present as example in the
\link{NBTest} man page.
}
}
\seealso{
Other NBSpliceRes: \code{\link{GetDSGenes}},
    \code{\link{GetDSResults}}, \code{\link{GetGeneResults}},
    \code{\link{NBSpliceRes-class}},
    \code{\link{NBSpliceRes-initialize}},
    \code{\link{NBSpliceRes}}, \code{\link{plotGeneResults}},
    \code{\link{plotRatiosDisp}}, \code{\link{plotVolcano}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and 
Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
