% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSplice-myIsoDataSet.R
\docType{data}
\name{myIsoDataSet}
\alias{myIsoDataSet}
\title{An example of an IsoDataSet object.}
\format{A data.frame object}
\source{
see \code{\link{IsoDataSet-class}}
}
\description{
An example of an IsoDataSet object helpful to demonstrate the use of
the NBSplice R package.
}
\details{
\describe{
An IsoDataSet object built with the package example dataset.
}
}
\seealso{
Other IsoDataSet: \code{\link{IsoDataSet-class}},
    \code{\link{IsoDataSet}}, \code{\link{NBTest}},
    \code{\link{buildData}}, \code{\link{buildLowExpIdx}},
    \code{\link{designMatrix}}, \code{\link{geneIso}},
    \code{\link{initialize}}, \code{\link{isoCountsData}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and 
Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
