% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSplice-package.R
\docType{package}
\name{NBSplice-package}
\alias{NBSplice-package}
\title{NBSplice: Negative Binomial Models to detect Differential Splicing.}
\description{
The package proposes a differential splicing evaluation method based on 
isoform quantification data. It uses generalized linear models with negative 
binomial distribution to model changes in isoform relative expression and
infer differential changes in gene splicing patterns.
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and 
Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
