% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes.R
\docType{class}
\name{NBSpliceRes-class}
\alias{NBSpliceRes-class}
\title{NBSpliceRes S4 class implementation in R}
\description{
This S4 class represents a data set containing NBTest results results.
}
\section{Slots}{

\describe{
\item{\code{results}}{Data.frame with NBTest results of expressed isoforms.}

\item{\code{lowExpIndex}}{Numeric indicating the positions of low expressed 
    isoforms.}

\item{\code{contrast}}{Character indicating the contrast used for NBTest.}

\item{\code{dispersion}}{Numeric model dispersions.}
}}

\section{Features}{

\enumerate{
    \item Explore differential splicing occurrence. 
    \item Explore isoform relative expression an its changes.
    \item Combine the results of both gene and isoform levels analysis.
}
}

\section{Functions}{

NBSpliceRes S4 class includes the following functions:
\describe{
\item{initialize}{Constructor of NBSpliceRes objects.} 
\item{results}{Gets the results slot.}
\item{contrast}{Gets the contrast slot.}
\item{lowExpIndex}{Gets the lowExpIdx slot.}
\item{disp}{Gets the dispersion slot.}
\item{print}{Shows a NBSpliceRes object.}
\item{show}{Shows a NBSpliceRes object.}
\item{GetDSGenes}{Returns the list of differentially spliced genes.}
\item{GetDSResults}{Returns the differential splicing results.}
\item{GetGeneResults}{Returns the NBSplice results for an specific gene.}
}
}

\examples{

data(myDSResults, package="NBSplice")
myResults<-results(myDSResults, filter=FALSE)
myLowExpIdx<-lowExpIndex(myDSResults)
myContrast<-contrast(myDSResults)

myNewDSResults<-NBSpliceRes(myResults, myLowExpIdx, myContrast)

##Getting differentially spliced genes
myDSGenes<-GetDSGenes(myDSResults)

##Getting the results for differentially spliced genes
myDSResultsDF<-GetDSResults(myDSResults)

##Getting the results for a particular differentially spliced gene
myResults<-results(myDSResults)

## Select the first gene
gene<-myResults[,"gene"][1]

myGeneResults<-GetGeneResults(myDSResults, gene)

}
\seealso{
Other NBSpliceRes: \code{\link{GetDSGenes}},
    \code{\link{GetDSResults}}, \code{\link{GetGeneResults}},
    \code{\link{NBSpliceRes-initialize}},
    \code{\link{NBSpliceRes}}, \code{\link{myDSResults}},
    \code{\link{plotGeneResults}},
    \code{\link{plotRatiosDisp}}, \code{\link{plotVolcano}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A.
Fernandez \email{efernandez@bdmg.com.ar}
}
