% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-constructor.R
\docType{methods}
\name{NBSpliceRes}
\alias{NBSpliceRes}
\alias{NBSpliceRes-methods}
\title{NBSpliceRes constructor}
\usage{
NBSpliceRes(results, lowExpIndex, contrast, dispersion)
}
\arguments{
\item{results}{Data.frame with NBTest results of expressed isoforms.}

\item{lowExpIndex}{Numeric indicating the positions of low expressed isoforms.}

\item{contrast}{Character indicating the contrast used for NBTest.}

\item{dispersion}{Numeric with the estimated gene dispersions.}
}
\value{
NBSpliceRes object.
}
\description{
\code{NBSpliceRes} creates an object to store the results of differential
expression analysis performed by the NBTest function.
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{

data(myDSResults, package="NBSplice")
myResults<-results(myDSResults, filter=FALSE)
myLowExpIdx<-lowExpIndex(myDSResults)
myContrast<-contrast(myDSResults)
myDispersion<-disp(myDSResults)
myNewDSResults<-NBSpliceRes(myResults, myLowExpIdx, myContrast, myDispersion)

}
\seealso{
\code{\link{NBSpliceRes-class}}

Other NBSpliceRes: \code{\link{GetDSGenes}},
  \code{\link{GetDSResults}}, \code{\link{GetGeneResults}},
  \code{\link{NBSpliceRes-class}},
  \code{\link{initialize}}, \code{\link{myDSResults}},
  \code{\link{plotGeneResults}},
  \code{\link{plotRatiosDisp}}, \code{\link{plotVolcano}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com}, and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
