% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-getDSGenes.R
\docType{methods}
\name{GetDSGenes}
\alias{GetDSGenes}
\alias{GetDSGenes-methods}
\alias{GetDSGenes}
\alias{GetDSGenes,NBSpliceRes-method}
\title{Get differentially spliced genes.}
\usage{
GetDSGenes(myNBRes, adjusted = TRUE, p.value = 0.05)

\S4method{GetDSGenes}{NBSpliceRes}(myNBRes, adjusted = TRUE, p.value = 0.05)
}
\arguments{
\item{myNBRes}{NBSpliceRes class object.}

\item{adjusted}{Logical indicating if adjusted p values should be used.}

\item{p.value}{Numeric value between 0 and 1 giving the required family-wise
error rate or false discovery rate.}
}
\value{
A character with the names of differentially spliced genes.
}
\description{
\code{GetDSGenes} returns the list of genes identified as differentially 
spliced.
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{
data(myDSResults, package="NBSplice")

myDSGenes<-GetDSGenes(myDSResults)
}
\seealso{
\code{\link{NBSpliceRes}}

Other NBSpliceRes: \code{\link{GetDSResults}},
  \code{\link{GetGeneResults}},
  \code{\link{NBSpliceRes-class}},
  \code{\link{NBSpliceRes}}, \code{\link{initialize}},
  \code{\link{myDSResults}}, \code{\link{plotGeneResults}},
  \code{\link{plotRatiosDisp}}, \code{\link{plotVolcano}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
