% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-getDSResults.R
\docType{methods}
\name{GetDSResults}
\alias{GetDSResults}
\alias{GetDSResults-methods}
\alias{GetDSResults}
\alias{GetDSResults,NBSpliceRes-method}
\title{Get differential expression results of significant genes.}
\usage{
GetDSResults(myNBRes, adjusted = TRUE, p.value = 0.05)

\S4method{GetDSResults}{NBSpliceRes}(myNBRes, adjusted = TRUE,
  p.value = 0.05)
}
\arguments{
\item{myNBRes}{NBSpliceRes class object.}

\item{adjusted}{Logical indicating if adjusted p values should be used.}

\item{p.value}{Numeric value between 0 and 1 giving the required family-wise
error rate or false discovery rate.}
}
\value{
data.frame with the results obtainted by means of the NBTest method
}
\description{
\code{GetDSResults} returns the results obtained for those genes identified 
as differentially spliced.
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{
data(myDSResults, package="NBSplice")

myDSResultsDF<-GetDSResults(myDSResults)

}
\seealso{
\code{\link{NBSpliceRes}}

Other NBSpliceRes: \code{\link{GetDSGenes}},
  \code{\link{GetGeneResults}},
  \code{\link{NBSpliceRes-class}},
  \code{\link{NBSpliceRes}}, \code{\link{initialize}},
  \code{\link{myDSResults}}, \code{\link{plotGeneResults}},
  \code{\link{plotRatiosDisp}}, \code{\link{plotVolcano}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
