% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-getGeneResults.R
\docType{methods}
\name{GetGeneResults}
\alias{GetGeneResults}
\alias{GetGeneResults-methods}
\alias{GetGeneResults}
\alias{GetGeneResults,NBSpliceRes-method}
\title{Get differentially spliced genes.}
\usage{
GetGeneResults(myNBRes, gene, filterLowExpIso = TRUE)

\S4method{GetGeneResults}{NBSpliceRes}(myNBRes, gene, filterLowExpIso = TRUE)
}
\arguments{
\item{myNBRes}{NBSpliceRes class object.}

\item{gene}{Character indicating the gene name.}

\item{filterLowExpIso}{Logical indicating if lower-expression isoforms should
be filtered out.}
}
\value{
Data.frame object with gene results.
}
\description{
\code{GetGeneResults} returns the results obtained for the specified gene.
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{
data(myDSResults, package="NBSplice")

myResults<-results(myDSResults)

## Select the first gene
gene<-myResults[,"gene"][1]

myGeneResults<-GetGeneResults(myDSResults, gene)

}
\seealso{
\code{\link{NBSpliceRes}}

Other NBSpliceRes: \code{\link{GetDSGenes}},
  \code{\link{GetDSResults}},
  \code{\link{NBSpliceRes-class}},
  \code{\link{NBSpliceRes}}, \code{\link{initialize}},
  \code{\link{myDSResults}}, \code{\link{plotGeneResults}},
  \code{\link{plotRatiosDisp}}, \code{\link{plotVolcano}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
