% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-getters.R
\docType{methods}
\name{NBSpliceRes-getters}
\alias{NBSpliceRes-getters}
\alias{results}
\alias{results-methods}
\alias{results}
\alias{results,IsoDataSet-method}
\alias{contrast}
\alias{contrast-methods}
\alias{contrast}
\alias{contrast,NBSpliceRes-method}
\alias{lowExpIndex}
\alias{lowExpIndex,NBSpliceRes-method}
\alias{disp}
\alias{disp-methods}
\alias{disp}
\alias{disp,NBSpliceRes-method}
\title{Getters for an NBSpliceRes object.}
\usage{
results(object, filter = TRUE)

\S4method{results}{NBSpliceRes}(object, filter = TRUE)

contrast(object)

\S4method{contrast}{NBSpliceRes}(object)

\S4method{lowExpIndex}{NBSpliceRes}(object)

disp(object)

\S4method{disp}{NBSpliceRes}(object)
}
\arguments{
\item{object}{NBSpliceRes class object.}

\item{filter}{Logical indicating if low expressed isoforms should be filtered
out.}
}
\value{
according to the call one of the following objects can be returned
\item{data.frame}{NBTest results.}
\item{numeric}{Index of low expressed isoforms.}
\item{character}{Experiment contrast.}
\item{numeric}{Estimated gene dispersion.}
}
\description{
Obtain information about NBSpliceRes slots, according to the given
function call.
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{
data(myDSResults, package="NBSplice")

## Getting results slot filtering the low expressed isoforms
myResults<-results(myDSResults)

## Getting results slot keeping the low expressed isoforms
myResults<-results(myDSResults, filter=FALSE)
## Getting the contrast slot
myContrast<-contrast(myDSResults)
## Getting the lowExpIndex slot
myLowExpIndex<-lowExpIndex(myDSResults)

## Getting the dispersion slot
myDispersion<-disp(myDSResults)
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A.
Fernandez \email{efernandez@bdmg.com.ar}
}
