% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-plotGeneResults.R
\docType{methods}
\name{plotGeneResults}
\alias{plotGeneResults}
\alias{plotGeneResults-methods}
\alias{plotGeneResults}
\alias{plotGeneResults,NBSpliceRes-method}
\title{Method to obtain isoform's relative expression barplot for an specific
gene.}
\usage{
plotGeneResults(myNBRes, gene, filterLowExpIso = TRUE,
    filterNotSignificant = TRUE, adjusted = TRUE, p.value = 0.05,
    group = TRUE)

\S4method{plotGeneResults}{NBSpliceRes}(myNBRes, gene, filterLowExpIso = TRUE,
    filterNotSignificant = TRUE, adjusted = TRUE, p.value = 0.05,
    group = TRUE)
}
\arguments{
\item{myNBRes}{NBSpliceRes class object.}

\item{gene}{Character indicating the gene name.}

\item{filterLowExpIso}{Logical indicating if lower-expression isoforms should
be filtered out.}

\item{filterNotSignificant}{Logical indicating if not significant 
isoforms should be filtered out.}

\item{adjusted}{Logical indicating if adjusted p values should be used.}

\item{p.value}{Numeric value between 0 and 1 giving the required family-wise
error rate or false discovery rate.}

\item{group}{Logical indicating if isoforms bars should be stacked or not}
}
\value{
A ggplot object.
}
\description{
\code{plotGeneResults} returns ggplot object which illustrates the isoform's 
relative expression in the two contrasted conditions.
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{
data(myDSResults, package="NBSplice")
gene<-results(myDSResults)[,"gene"][1]
##Plot gene results filtering low expressed isoforms
g<-plotGeneResults(myDSResults, gene)
if(interactive()){
g
}
##Plot gene results keeping low expressed isoforms
g<-plotGeneResults(myDSResults, gene, filterLowExpIso=FALSE)
if(interactive()){
g
}
##Plot isoform bar plots keeping low expressed isoforms
g<-plotGeneResults(myDSResults, gene, filterLowExpIso=FALSE, group=FALSE)
if(interactive()){
g
}
}
\seealso{
\code{\link{NBSpliceRes}}

Other NBSpliceRes: \code{\link{GetDSGenes}},
    \code{\link{GetDSResults}}, \code{\link{GetGeneResults}},
    \code{\link{NBSpliceRes-class}},
    \code{\link{NBSpliceRes-initialize}},
    \code{\link{NBSpliceRes}}, \code{\link{myDSResults}},
    \code{\link{plotRatiosDisp}}, \code{\link{plotVolcano}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
