% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-plotVolcano.R
\docType{methods}
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano-methods}
\alias{plotVolcano}
\alias{plotVolcano,NBSpliceRes-method}
\title{Method to obtain the Volcano Plot.}
\usage{
plotVolcano(myNBRes, adjusted = TRUE, p.value = 0.05)

\S4method{plotVolcano}{NBSpliceRes}(myNBRes, adjusted = TRUE,
  p.value = 0.05)
}
\arguments{
\item{myNBRes}{NBSpliceRes class object.}

\item{adjusted}{Logical indicating if adjusted p values should be used.}

\item{p.value}{Numeric value between 0 and 1 giving the required family-wise
error rate or false discovery rate.}
}
\value{
A ggplot object.
}
\description{
\code{plotVolcano} returns a generalized volcano plot where the x-axis
represents the difference between isoform's relative expression in the 
contrasted conditions. Isoforms are colored according to the differential
splicing status of the gene where they come from.
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{
data(myDSResults, package="NBSplice")

g<-plotVolcano(myDSResults)
if(interactive()){
g
}
}
\seealso{
\code{\link{NBSpliceRes}}

Other NBSpliceRes: \code{\link{GetDSGenes}},
  \code{\link{GetDSResults}}, \code{\link{GetGeneResults}},
  \code{\link{NBSpliceRes-class}},
  \code{\link{NBSpliceRes}}, \code{\link{initialize}},
  \code{\link{myDSResults}}, \code{\link{plotGeneResults}},
  \code{\link{plotRatiosDisp}}
}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A. 
Fernandez \email{efernandez@bdmg.com.ar}
}
