% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBSpliceRes-show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,NBSpliceRes-method}
\title{Show method for the NBSpliceRes class.}
\usage{
\S4method{show}{NBSpliceRes}(object)
}
\arguments{
\item{object}{NBSpliceRes class object}
}
\value{
Console output of the object
}
\description{
\code{show} an NBSpliceRes object
}
\note{
see full example in \code{\link{NBSpliceRes-class}}
}
\examples{

data(myDSResults, package="NBSplice")

show(myDSResults)

}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com}
and Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
