% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGeneCounts.R
\docType{methods}
\name{totalGeneCounts}
\alias{totalGeneCounts}
\alias{totalGeneCounts-methods}
\title{Auxiliary function to build gene expression matrix}
\usage{
totalGeneCounts(isoCounts, geneIso, BPPARAM = bpparam())
}
\arguments{
\item{isoCounts}{Matrix containing the isoform expression values with
isoforms in rows and samples in columns. Object row names should 
be the names of the isoforms and the column names, the names of each sample.}

\item{geneIso}{Data.frame with two columns called "isoform_id" and "gene_id"
especifying which are the isoform (rows of the expression matrix) for each
gene}

\item{BPPARAM}{A bpparam object specifying the number of cpus to be used.}
}
\value{
A matrix of the same size as the isoCounts matrix with the total 
counts for each gene in each sample in CPM scale.
}
\description{
\code{totalGeneCounts} obtains the total counts for a gene in a sample by the
sum of the counts of all its isoforms. This count is performed before any
filter step to avoid erroneus estimation of isoform ratios.
}
\examples{

data(isoCounts, package="NBSplice")
data(geneIso, package="NBSplice")

myGeneCounts<-totalGeneCounts(isoCounts, geneIso)

}
\author{
Gabriela A. Merino \email{merino.gabriela33@gmail.com} and Elmer A.
Fernandez \email{efernandez@bdmg.com.ar}
}
